@@ -36,18 +36,18 @@ repositories {
 }
 
 java {
-    sourceCompatibility = JavaVersion.VERSION_1_8
-    targetCompatibility = JavaVersion.VERSION_1_8
+    sourceCompatibility = JavaVersion.VERSION_11
+    targetCompatibility = JavaVersion.VERSION_11
 }
 
 checkstyle {
-    toolVersion "8.19"
+    toolVersion "8.35"
     configDir rootProject.file("checkstyle/")
     println configFile
 }
 
 jacoco {
-    toolVersion = "0.8.3"
+    toolVersion = "0.8.5"
 }
 
 distributions {
@@ -58,7 +58,7 @@ distributions {
 
             into("/") {
                 from projectDir
-                include "version.txt", "README*", "LICENSE*", "NOTICE*", "licenses/"
+                include "README*", "LICENSE*", "NOTICE*", "licenses/"
                 include "config/"
             }
         }
@@ -108,9 +108,12 @@ ext {
 dependencies {
     compileOnly "org.apache.kafka:connect-api:$kafkaVersion"
 
-    runtimeOnly "org.xerial:sqlite-jdbc:3.25.2"
-    runtimeOnly "org.postgresql:postgresql:42.2.9"
+    runtimeOnly "org.xerial:sqlite-jdbc:3.32.3"
+    runtimeOnly "org.postgresql:postgresql:42.2.10"
     runtimeOnly "net.sourceforge.jtds:jtds:1.3.1"
+    runtimeOnly "net.snowflake:snowflake-jdbc:3.12.8"
+    runtimeOnly "com.microsoft.sqlserver:mssql-jdbc:8.2.1.jre11"
+    runtimeOnly "mysql:mysql-connector-java:8.0.21"
 
     implementation "com.google.guava:guava:27.1-jre"
     implementation "org.slf4j:slf4j-api:$slf4jVersion"
@@ -124,6 +127,7 @@ dependencies {
     testImplementation "commons-io:commons-io:2.6"
     testImplementation "org.apache.derby:derby:10.11.1.1"
     testRuntime "org.slf4j:slf4j-log4j12:$slf4jVersion"
+
 }
 
 processResources {
