@@ -43,6 +43,7 @@ import io.aiven.connect.jdbc.util.ColumnId;
 import io.aiven.connect.jdbc.util.ExpressionBuilder;
 import io.aiven.connect.jdbc.util.IdentifierRules;
 import io.aiven.connect.jdbc.util.TableId;
+import io.aiven.connect.jdbc.util.TableDefinition;
 
 
 /**
@@ -238,21 +239,23 @@ public class RedshiftDatabaseDialect extends GenericDatabaseDialect {
     @Override
     public String buildInsertStatement(
             final TableId table,
+            final TableDefinition tableDefinition,
             final Collection<ColumnId> keyColumns,
             final Collection<ColumnId> nonKeyColumns
     ) {
-        final ExpressionBuilder builder = expressionBuilder();
-        builder.append("INSERT INTO ");
-        builder.append(table);
-        builder.append("(");
-        builder.appendList()
+        log.debug("redshift: building insert statement for table {}", table.tableName());
+        return expressionBuilder()
+                .append("INSERT INTO ")
+                .append(table)
+                .append("(")
+                .appendList()
                 .delimitedBy(",")
                 .transformedBy(ExpressionBuilder.columnNames())
-                .of(keyColumns, nonKeyColumns);
-        builder.append(") VALUES(");
-        builder.appendMultiple(",", "?", keyColumns.size() + nonKeyColumns.size());
-        builder.append(")");
-        return builder.toString();
+                .of(keyColumns, nonKeyColumns)
+                .append(") VALUES(")
+                .appendMultiple(",", "?", keyColumns.size() + nonKeyColumns.size())
+                .append(")")
+                .toString();
     }
 
     @Override
@@ -261,18 +264,19 @@ public class RedshiftDatabaseDialect extends GenericDatabaseDialect {
             final Collection<ColumnId> keyColumns,
             final Collection<ColumnId> nonKeyColumns
     ) {
-        final ExpressionBuilder builder = expressionBuilder();
-        builder.append("INSERT INTO ");
-        builder.append(table);
-        builder.append("(");
-        builder.appendList()
+        log.debug("redshift: building first multi insert statement for table {}", table.tableName());
+        return expressionBuilder()
+                .append("INSERT INTO ")
+                .append(table)
+                .append("(")
+                .appendList()
                 .delimitedBy(",")
                 .transformedBy(ExpressionBuilder.columnNames())
-                .of(keyColumns, nonKeyColumns);
-        builder.append(") VALUES(");
-        builder.appendMultiple(",", "?", keyColumns.size() + nonKeyColumns.size());
-        builder.append(")");
-        return builder.toString();
+                .of(keyColumns, nonKeyColumns)
+                .append(") VALUES(")
+                .appendMultiple(",", "?", keyColumns.size() + nonKeyColumns.size())
+                .append(")")
+                .toString();
     }
 
     @Override
@@ -281,10 +285,11 @@ public class RedshiftDatabaseDialect extends GenericDatabaseDialect {
             final Collection<ColumnId> keyColumns,
             final Collection<ColumnId> nonKeyColumns
     ) {
-        final ExpressionBuilder builder = expressionBuilder();
-        builder.append(",(");
-        builder.appendMultiple(",", "?", keyColumns.size() + nonKeyColumns.size());
-        builder.append(")");
-        return builder.toString();
+        log.debug("redshift: building multi insert statement for table {}", table.tableName());
+        return expressionBuilder()
+            .append(",(")
+            .appendMultiple(",", "?", keyColumns.size() + nonKeyColumns.size())
+            .append(")")
+            .toString();
     }
 }
