@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Aiven Oy
+ * Copyright 2020 Aiven Oy
  * Copyright 2016 Confluent Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -23,6 +23,7 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Objects;
 import java.util.Set;
 
 import org.apache.kafka.connect.errors.ConnectException;
@@ -48,15 +49,24 @@ public class DbStructure {
         this.tableDefns = new TableDefinitions(dbDialect);
     }
 
+    public TableDefinition tableDefinitionFor(final TableId tableId, final Connection connection) throws SQLException {
+        final var tblDefinition = tableDefns.get(connection, tableId);
+        if (Objects.nonNull(tblDefinition)) {
+            return tblDefinition;
+        } else {
+            return tableDefns.refresh(connection, tableId);
+        }
+    }
+
     /**
      * @return whether a DDL operation was performed
      * @throws SQLException if a DDL operation was deemed necessary but failed
      */
     public boolean createOrAmendIfNecessary(
-        final JdbcSinkConfig config,
-        final Connection connection,
-        final TableId tableId,
-        final FieldsMetadata fieldsMetadata
+            final JdbcSinkConfig config,
+            final Connection connection,
+            final TableId tableId,
+            final FieldsMetadata fieldsMetadata
     ) throws SQLException {
         if (tableDefns.get(connection, tableId) == null) {
             // Table does not yet exist, so attempt to create it ...
@@ -81,14 +91,14 @@ public class DbStructure {
      * @throws SQLException if CREATE failed
      */
     void create(
-        final JdbcSinkConfig config,
-        final Connection connection,
-        final TableId tableId,
-        final FieldsMetadata fieldsMetadata
+            final JdbcSinkConfig config,
+            final Connection connection,
+            final TableId tableId,
+            final FieldsMetadata fieldsMetadata
     ) throws SQLException {
         if (!config.autoCreate) {
             throw new ConnectException(
-                String.format("Table %s is missing and auto-creation is disabled", tableId)
+                    String.format("Table %s is missing and auto-creation is disabled", tableId)
             );
         }
         final String sql = dbDialect.buildCreateTableStatement(tableId, fieldsMetadata.allFields.values());
@@ -101,11 +111,11 @@ public class DbStructure {
      * @throws SQLException if ALTER was deemed necessary but failed
      */
     boolean amendIfNecessary(
-        final JdbcSinkConfig config,
-        final Connection connection,
-        final TableId tableId,
-        final FieldsMetadata fieldsMetadata,
-        final int maxRetries
+            final JdbcSinkConfig config,
+            final Connection connection,
+            final TableId tableId,
+            final FieldsMetadata fieldsMetadata,
+            final int maxRetries
     ) throws SQLException {
         // NOTE:
         //   The table might have extra columns defined (hopefully with default values), which is not
@@ -123,8 +133,8 @@ public class DbStructure {
         //    }
 
         final Set<SinkRecordField> missingFields = missingFields(
-            fieldsMetadata.allFields.values(),
-            tableDefn.columnNames()
+                fieldsMetadata.allFields.values(),
+                tableDefn.columnNames()
         );
 
         if (missingFields.isEmpty()) {
@@ -134,49 +144,49 @@ public class DbStructure {
         for (final SinkRecordField missingField : missingFields) {
             if (!missingField.isOptional() && missingField.defaultValue() == null) {
                 throw new ConnectException(
-                    "Cannot ALTER to add missing field " + missingField
-                        + ", as it is not optional and does not have a default value"
+                        "Cannot ALTER to add missing field " + missingField
+                                + ", as it is not optional and does not have a default value"
                 );
             }
         }
 
         if (!config.autoEvolve) {
             throw new ConnectException(String.format(
-                "Table %s is missing fields (%s) and auto-evolution is disabled",
-                tableId,
-                missingFields
+                    "Table %s is missing fields (%s) and auto-evolution is disabled",
+                    tableId,
+                    missingFields
             ));
         }
 
         final List<String> amendTableQueries = dbDialect.buildAlterTable(tableId, missingFields);
         log.info(
-            "Amending table to add missing fields:{} maxRetries:{} with SQL: {}",
-            missingFields,
-            maxRetries,
-            amendTableQueries
+                "Amending table to add missing fields:{} maxRetries:{} with SQL: {}",
+                missingFields,
+                maxRetries,
+                amendTableQueries
         );
         try {
             dbDialect.applyDdlStatements(connection, amendTableQueries);
         } catch (final SQLException sqle) {
             if (maxRetries <= 0) {
                 throw new ConnectException(
-                    String.format(
-                        "Failed to amend table '%s' to add missing fields: %s",
-                        tableId,
-                        missingFields
-                    ),
-                    sqle
+                        String.format(
+                                "Failed to amend table '%s' to add missing fields: %s",
+                                tableId,
+                                missingFields
+                        ),
+                        sqle
                 );
             }
             log.warn("Amend failed, re-attempting", sqle);
             tableDefns.refresh(connection, tableId);
             // Perhaps there was a race with other tasks to add the columns
             return amendIfNecessary(
-                config,
-                connection,
-                tableId,
-                fieldsMetadata,
-                maxRetries - 1
+                    config,
+                    connection,
+                    tableId,
+                    fieldsMetadata,
+                    maxRetries - 1
             );
         }
 
@@ -185,8 +195,8 @@ public class DbStructure {
     }
 
     Set<SinkRecordField> missingFields(
-        final Collection<SinkRecordField> fields,
-        final Set<String> dbColumnNames
+            final Collection<SinkRecordField> fields,
+            final Set<String> dbColumnNames
     ) {
         final Set<SinkRecordField> missingFields = new HashSet<>();
         for (final SinkRecordField field : fields) {
@@ -208,8 +218,8 @@ public class DbStructure {
 
         if (columnNamesLowerCase.size() != dbColumnNames.size()) {
             log.warn(
-                "Table has column names that differ only by case. Original columns={}",
-                dbColumnNames
+                    "Table has column names that differ only by case. Original columns={}",
+                    dbColumnNames
             );
         }
 
@@ -222,9 +232,9 @@ public class DbStructure {
 
         if (missingFieldsIgnoreCase.size() > 0) {
             log.info(
-                "Unable to find fields {} among column names {}",
-                missingFieldsIgnoreCase,
-                dbColumnNames
+                    "Unable to find fields {} among column names {}",
+                    missingFieldsIgnoreCase,
+                    dbColumnNames
             );
         }
 
