@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Aiven Oy
+ * Copyright 2020 Aiven Oy
  * Copyright 2016 Confluent Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -24,12 +24,13 @@ import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Set;
 import java.util.TimeZone;
+import java.util.stream.Collectors;
 
 import org.apache.kafka.common.config.ConfigDef;
 import org.apache.kafka.common.config.ConfigException;
-import org.apache.kafka.common.config.types.Password;
 
 import io.aiven.connect.jdbc.config.JdbcConfig;
 import io.aiven.connect.jdbc.util.StringUtils;
@@ -59,13 +60,29 @@ public class JdbcSinkConfig extends JdbcConfig {
     );
 
     public static final String TABLE_NAME_FORMAT = "table.name.format";
-    private static final String TABLE_NAME_FORMAT_DEFAULT = "${topic}";
+    public static final String TABLE_NAME_FORMAT_DEFAULT = "${topic}";
     private static final String TABLE_NAME_FORMAT_DOC =
         "A format string for the destination table name, which may contain '${topic}' as a "
             + "placeholder for the originating topic name.\n"
             + "For example, ``kafka_${topic}`` for the topic 'orders' will map to the table name "
             + "'kafka_orders'.";
     private static final String TABLE_NAME_FORMAT_DISPLAY = "Table Name Format";
+
+    public static final String TABLE_NAME_NORMALIZE = "table.name.normalize";
+    public static final boolean TABLE_NAME_NORMALIZE_DEFAULT = false;
+    private static final String TABLE_NAME_NORMALIZE_DOC =
+            "Whether or not to normalize destination table names for topics. "
+                    + "When set to ``true``, the alphanumeric characters (``a-z A-Z 0-9``) and ``_`` "
+                    + "remain as is, others (like ``.``) are replaced with ``_``.";
+    private static final String TABLE_NAME_NORMALIZE_DISPLAY = "Table Name Normalize";
+
+    public static final String TOPICS_TO_TABLES_MAPPING = "topics.to.tables.mapping";
+    private static final String TOPICS_TO_TABLES_MAPPING_DOC =
+            "Kafka topics to database tables mapping. "
+                    + "Comma-separated list of topic to table mapping in the format: topic_name:table_name. "
+                    + "If the destination table found in the mapping, "
+                    + "it would override generated one defined in " + TABLE_NAME_FORMAT + ".";
+    private static final String TOPICS_TO_TABLES_MAPPING_DISPLAY = "Topics To Tables Mapping";
 
     public static final String MAX_RETRIES = "max.retries";
     private static final int MAX_RETRIES_DEFAULT = 10;
@@ -217,6 +234,47 @@ public class JdbcSinkConfig extends JdbcConfig {
                 TABLE_NAME_FORMAT_DISPLAY
             )
             .define(
+                TABLE_NAME_NORMALIZE,
+                ConfigDef.Type.BOOLEAN,
+                TABLE_NAME_NORMALIZE_DEFAULT,
+                ConfigDef.Importance.MEDIUM,
+                TABLE_NAME_NORMALIZE_DOC,
+                DATAMAPPING_GROUP,
+                2,
+                ConfigDef.Width.LONG,
+                TABLE_NAME_NORMALIZE_DISPLAY
+            )
+            .define(
+                TOPICS_TO_TABLES_MAPPING,
+                ConfigDef.Type.LIST,
+                null,
+                new ConfigDef.Validator() {
+                    @Override
+                    public void ensureValid(final String name, final Object value) {
+                        if (Objects.isNull(value)
+                                || ConfigDef.NO_DEFAULT_VALUE == value
+                                || "".equals(value)) {
+                            return;
+                        }
+                        assert value instanceof List;
+                        try {
+                            final Map<String, String> mapping = topicToTableMapping((List<String>) value);
+                            if (Objects.isNull(mapping) || mapping.isEmpty()) {
+                                throw new ConfigException(name, value, "Invalid topics to tables mapping");
+                            }
+                        } catch (final ArrayIndexOutOfBoundsException e) {
+                            throw new ConfigException(name, value, "Invalid topics to tables mapping");
+                        }
+                    }
+                },
+                ConfigDef.Importance.MEDIUM,
+                TOPICS_TO_TABLES_MAPPING_DOC,
+                DATAMAPPING_GROUP,
+                3,
+                ConfigDef.Width.LONG,
+                TOPICS_TO_TABLES_MAPPING_DISPLAY
+            )
+            .define(
                 PK_MODE,
                 ConfigDef.Type.STRING,
                 PK_MODE_DEFAULT,
@@ -224,7 +282,7 @@ public class JdbcSinkConfig extends JdbcConfig {
                 ConfigDef.Importance.HIGH,
                 PK_MODE_DOC,
                 DATAMAPPING_GROUP,
-                2,
+                4,
                 ConfigDef.Width.MEDIUM,
                 PK_MODE_DISPLAY
             )
@@ -235,7 +293,7 @@ public class JdbcSinkConfig extends JdbcConfig {
                 ConfigDef.Importance.MEDIUM,
                 PK_FIELDS_DOC,
                 DATAMAPPING_GROUP,
-                3,
+                5,
                 ConfigDef.Width.LONG, PK_FIELDS_DISPLAY
             )
             .define(
@@ -245,7 +303,7 @@ public class JdbcSinkConfig extends JdbcConfig {
                 ConfigDef.Importance.MEDIUM,
                 FIELDS_WHITELIST_DOC,
                 DATAMAPPING_GROUP,
-                4,
+                6,
                 ConfigDef.Width.LONG,
                 FIELDS_WHITELIST_DISPLAY
             );
@@ -300,6 +358,8 @@ public class JdbcSinkConfig extends JdbcConfig {
     }
 
     public final String tableNameFormat;
+    public final Map<String, String> topicsToTablesMapping;
+    public final boolean tableNameNormalize;
     public final int batchSize;
     public final int maxRetries;
     public final int retryBackoffMs;
@@ -314,6 +374,8 @@ public class JdbcSinkConfig extends JdbcConfig {
     public JdbcSinkConfig(final Map<?, ?> props) {
         super(CONFIG_DEF, props);
         tableNameFormat = getString(TABLE_NAME_FORMAT).trim();
+        tableNameNormalize = getBoolean(TABLE_NAME_NORMALIZE);
+        topicsToTablesMapping = topicToTableMapping(getList(TOPICS_TO_TABLES_MAPPING));
         batchSize = getInt(BATCH_SIZE);
         maxRetries = getInt(MAX_RETRIES);
         retryBackoffMs = getInt(RETRY_BACKOFF_MS);
@@ -327,12 +389,12 @@ public class JdbcSinkConfig extends JdbcConfig {
         timeZone = TimeZone.getTimeZone(ZoneId.of(dbTimeZone));
     }
 
-    private String getPasswordValue(final String key) {
-        final Password password = getPassword(key);
-        if (password != null) {
-            return password.value();
-        }
-        return null;
+    static Map<String, String> topicToTableMapping(final List<String> value) {
+        return (Objects.nonNull(value))
+                ? value.stream()
+                    .map(s -> s.split(":"))
+                    .collect(Collectors.toMap(e -> e[0], e -> e[1]))
+                : Collections.emptyMap();
     }
 
     private static class EnumValidator implements ConfigDef.Validator {
